EgoPlan-challenge-Team-AAILab/requirements.txt
absl-py==2.1.0
accelerate==0.24.1
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
av==11.0.0
Babel==2.15.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bitsandbytes==0.42.0
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
braceexpand==0.1.7
Brotli==1.1.0
cachetools==4.2.2
catalogue==2.0.10
certifi==2021.5.30
cffi==1.16.0
cfgv==3.4.0
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cloudpathlib==0.16.0
cmake==3.26.4
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.3
contexttimer==0.3.3
cycler==0.10.0
cymem==2.0.8
debugpy==1.8.1
decorator==5.1.1
decord==0.6.0
deepspeed==0.14.2
defusedxml==0.7.1
distlib==0.3.7
distro==1.9.0
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.7.0
en-core-web-sm==3.7.1
exceptiongroup==1.1.1
executing==2.0.1
fairscale==0.4.4
fastapi==0.98.0
fastjsonschema==2.20.0
ffmpeg-python==0.2.0
ffmpy==0.3.0
filelock==3.13.1
fire==0.5.0
flatbuffers==24.3.25
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
ftfy==6.1.1
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.4.0
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.30.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
GPUtil==1.4.0
gradio==3.35.2
gradio_client==0.2.7
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
hjson==3.0.2
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
identify==2.5.31
idna==2.10
imageio==2.33.0
importlib-metadata==6.8.0
importlib-resources==5.12.0
install==1.3.5
iopath==0.1.10
ipykernel==6.29.4
ipython==8.12.3
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kaggle==1.5.16
keras==2.13.1
kiwisolver==1.3.1
langcodes==3.3.0
langdetect==1.0.9
lazy_loader==0.3
libclang==18.1.1
linkify-it-py==2.0.2
lit==16.0.6
lmdb==1.2.1
loguru==0.5.3
Markdown==3.3.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.4.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mutagen==1.47.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
ninja==1.10.2.3
nltk==3.8.1
nodeenv==1.8.0
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.1
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.18.0
openai==1.30.1
opencv-python==4.8.0.74
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
opt-einsum==3.3.0
orjson==3.9.1
overrides==7.7.0
packaging==24.1
pandas==1.3.5
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.3
peft==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pipreqs==0.4.13
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
plotly==5.18.0
portalocker==2.8.2
pre-commit==3.5.0
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.8.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.22
pycryptodomex==3.20.0
pydantic==1.10.9
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.15.1
pynvml==11.5.0
pyparsing==2.4.7
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
python-slugify==8.0.1
pytorchvideo==0.1.5
pytz==2021.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.18.1
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.3.1
salesforce-lavis==1.0.2
scenedetect==0.6.2
scikit-image==0.21.0
scipy==1.10.1
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.1.99
sentry-sdk==2.3.1
setproctitle==1.3.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
streamlit==1.28.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
terminado==0.18.1
text-unidecode==1.3
tf-keras==2.15.0
thinc==8.2.1
tifffile==2023.7.10
timeout-decorator==0.5.0
timm==0.4.12
tinycss2==1.3.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchdata==0.7.1
torchvision==0.14.1+cu117
tornado==6.3.3
tqdm==4.61.1
traitlets==5.13.0
transformers==4.37.1
triton==2.1.0
typer==0.9.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.5.0
tzlocal==5.2
uc-micro-py==1.0.2
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.22.0
validators==0.22.0
video2dataset==1.3.0
virtualenv==20.24.6
wandb==0.17.0
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.9
weasel==0.3.4
webcolors==24.6.0
webdataset==0.2.73
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
webvtt-py==0.5.0
Werkzeug==2.0.1
widgetsnbextension==4.0.11
wordcloud==1.9.3
wrapt==1.16.0
yacs==0.1.8
yarg==0.1.9
yarl==1.9.2
yt-dlp==2024.4.9
zipp==3.15.0